src/app/katwalk/requirements.txt
vllm
torch
uvicorn
fastapi
gitpython